Checkpoint of encapsulating the loading stuff
